+++ thesis/Charter/PlotFolder/LinePlot.java
@@ -112,11 +112,11 @@ public class LinePlot extends Plot{
 	}
 	
 	private double xPlotValueToXPixelValue(double xPos, double[] xTicks, XYChartMeasurements cm) {
-		return (int) CommonMath.map(xPos, xTicks[0], xTicks[xTicks.length - 1], cm.imageLeftToPlotLeftWidth(), cm.imageLeftToPlotRightWidth());
+		return CommonMath.map(xPos, xTicks[0], xTicks[xTicks.length - 1], cm.imageLeftToPlotLeftWidth(), cm.imageLeftToPlotRightWidth());
 	}
 	
 	private double yPlotValueToYPixelValue(double yPos, double[] yTicks, XYChartMeasurements cm) {
-		return (int) CommonMath.map(yPos, yTicks[0], yTicks[yTicks.length - 1 ], cm.imageBottomToPlotBottomHeight(), cm.imageBottomToPlotTopHeight());
+		return CommonMath.map(yPos, yTicks[0], yTicks[yTicks.length - 1 ], cm.imageBottomToPlotBottomHeight(), cm.imageBottomToPlotTopHeight());
 	}
 	
 	public void setLineColor(Color lineColor) {
