+++ thesis/Charter/PlotFolder/ScatterPlot.java
@@ -173,11 +173,11 @@ public class ScatterPlot extends Plot {
 	}
 	
 	private double worldXPosToPlotXPos(double xPos, NiceScale xNS, double[] xTicks, XYChartMeasurements cm) {
-		return (int) CommonMath.map(xPos, xNS.getNiceMin(), xTicks[xTicks.length - 1], cm.imageLeftToPlotLeftWidth(), cm.imageLeftToPlotRightWidth());
+		return CommonMath.map(xPos, xNS.getNiceMin(), xTicks[xTicks.length - 1], cm.imageLeftToPlotLeftWidth(), cm.imageLeftToPlotRightWidth());
 	}
 	
 	private double worldYPosToPlotYPos(double yPos, NiceScale yNS, double[] yTicks, XYChartMeasurements cm) {
-		return (int) CommonMath.map(yPos, yNS.getNiceMin(), yTicks[yTicks.length - 1 ], cm.imageBottomToPlotBottomHeight(), cm.imageBottomToPlotTopHeight());
+		return CommonMath.map(yPos, yNS.getNiceMin(), yTicks[yTicks.length - 1 ], cm.imageBottomToPlotBottomHeight(), cm.imageBottomToPlotTopHeight());
 	}
